requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
boto3==1.42.78
botocore==1.42.78
bs4==0.0.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
executing==2.2.1
faiss-cpu==1.13.2
fastapi==0.135.2
filelock==3.25.2
from-root==1.3.0
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.2
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.22
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
logger==1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
-e git+https://github.com/lalit7886/Multi_Modal_Search_Engine.git@a021011ab3292fe3a16bfa070e0baf889b5fe009#egg=multimodal_search_engine
mypy-boto3-s3==1.42.67
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.3
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
parso==0.8.5
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.5.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.2.28
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
s3transfer==0.16.0
safetensors==0.7.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
-e git+https://github.com/lalit7886/Multi_Modal_Search_Engine.git@a021011ab3292fe3a16bfa070e0baf889b5fe009#egg=src
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
wcwidth==0.5.0
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
